bot/requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
openpyxl==3.1.2
pip==24.0
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
typing_extensions==4.11.0
yarl==1.9.4
motor==3.4.0